Refactor rewrite rules into the rewriter.rules namespace
Organize all rules into a directory that is not with the rewriter infrastructure
