+++ onnxscript/rewriter/rules/cast_constant_of_shape.py
