+++ onnxscript/rewriter/rules/collapse_slices.py
