+++ onnxscript/rewriter/rules/fuse_relus_clips_test.py
@@ -13,14 +13,14 @@ from onnxscript.rewriter import (
     MatchingTracer,
     MatchStatus,
     RewriteRule,
-    fuse_relus_clips,
     testing,
 )
-from onnxscript.rewriter.fuse_relus_clips import (
+from onnxscript.rewriter.rules.fuse_relus_clips import (
     fuse_successive_clip_relu_rule,
     fuse_successive_clip_rule,
     fuse_successive_relu_clip_rule,
 )
+from onnxscript.rewriter.rules import fuse_relus_clips
 
 
 class _FuseReluClipTestBase(unittest.TestCase):
