+++ onnxscript/rewriter/rules/no_op.py
