+++ onnxscript/rewriter/rules/redundant_scatter_nd.py
